Сплитит gulp таски
