+++ gulp/compileStyles.mjs
@@ -1,0 +1,26 @@
+import gulp from 'gulp';
+import plumber from 'gulp-plumber';
+import dartSass from 'sass';
+import gulpSass from 'gulp-sass';
+import postcss from 'gulp-postcss';
+import autoprefixer from 'autoprefixer';
+import csso from 'gulp-csso';
+import gcmq from 'gulp-group-css-media-queries';
+import rename from 'gulp-rename';
+
+const sass = gulpSass(dartSass);
+
+const compileStyles = () =>
+  gulp.src('source/sass/style.scss', {sourcemaps: true})
+      .pipe(plumber())
+      .pipe(sass())
+      .pipe(postcss([autoprefixer({
+        grid: true,
+      })]))
+      .pipe(gcmq()) // выключите, если в проект импортятся шрифты через ссылку на внешний источник
+      .pipe(gulp.dest('build/css'))
+      .pipe(csso())
+      .pipe(rename('style.min.css'))
+      .pipe(gulp.dest('build/css', {sourcemaps: '.'}));
+
+export default compileStyles;
